crouching and making weapon and attack BPs more modular
Added crouching and lock on. Modified the weapon BP base to make it so that a weapon is modular the attack system. It needs a mesh, light, heavy and special attack animations. Can have up to 5 of each to allow for combos.  LS basic is a good template for how to structure a weapon. Any new weapon anim will need to get a montages made and need have anim notifies added, again see LS basic as a general guide.