Began AI combat implimentation
Combat feels actually pretty good for how early on it is in development. The enemy is way to aggressive and needs to take a break between combos because its a mad man right now. I need to implement a way to stun them, something like posture breaking. I also need to add a healing item. But I think once I get the combat in a place where I like it, I will start working on a boss enemy design.